// src/models/institutions/TDAmeritradeRecord.ts
export interface TDAmeritradeRecord {
  DATE: string;
  'TRANSACTION ID': string;
  DESCRIPTION: string;
  QUANTITY: string;
  SYMBOL: string;
  PRICE: string;
  COMMISSION: string;
  AMOUNT: string;
  'NET CASH BALANCE': string;
  'REG FEE': string;
  'SHORT-TERM RDM FEE': string;
  'FUND REDEMPTION FEE': string;
  'DEFERRED SALES CHARGE': string;
}